+++ db/models/Task.js
@@ -16,8 +16,8 @@ const taskSchema = new Schema({
     default: [],
   },
   priority: { type: String },
-  originalTaskDescription: { type: String },
-  creationDate: { type: Date },
+  original_task_description: { type: String },
+  creation_date: { type: Date },
 });
 
 const Task = mongoose.models.Task || mongoose.model("Task", taskSchema);
