+++ pages/tasks/[dynamicId]/edit.js
@@ -1,27 +1,27 @@
 import { useRouter } from "next/router";
 import useSWR from "swr";
-import Form from "../../../components/TaskInputForm";
 import Link from "next/link";
 import styled from "styled-components";
+import RegularTaskInputForm from "../../../components/RegularTaskInputForm";
 
 export default function TaskEditPage() {
   const router = useRouter();
   const { isReady } = router;
   const { dynamicId } = router.query;
   const {
-    data: task,
+    data: existingTaskData,
     isLoading,
     error,
     mutate,
   } = useSWR(`/api/tasks/${dynamicId}`);
 
-  async function editTask(task) {
+  async function editTask(existingTaskData) {
     const response = await fetch(`/api/tasks/${dynamicId}`, {
       method: "PATCH",
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(task),
+      body: JSON.stringify(existingTaskData),
     });
     if (response.ok) {
       mutate();
@@ -33,10 +33,10 @@ export default function TaskEditPage() {
 
   return (
     <>
-      <Form
+      <RegularTaskInputForm
         onSubmit={editTask}
         formName={"edit-task"}
-        existingTaskData={task}
+        existingTaskData={existingTaskData}
       />
       <LinkWrapper>
         <Link href={`/`} passHref legacyBehavior aria-label="go back">
